+++ tests/SPSUserSync.Common.Tests.ps1
@@ -23,6 +23,7 @@ Describe 'SPSUserSync.Common module' {
             'Backup-SPSJsonFile'
             'Clear-SPSLogFolder'
             'Compare-SPSJsonSnapshots'
+            'ConvertTo-SPSUserRecord'
             'Export-SPSUserReport'
             'Get-SPSADConnection'
             'Get-SPSADUser'